implement a generic unix NTP clock (for freebsd, macos)
our existing implementation uses linux-specific functionality to reduce the number of syscalls. This is not possible in general on all unix systems
